# parse: speed improved