**faster by 15%**

before:

```console
       4,644,092 op/s » block
       3,235,655 op/s » blockMod
       3,221,562 op/s » elem
       2,210,935 op/s » elemMod
```

after:

```console
       6,485,415 op/s » block
       3,358,493 op/s » blockMod
       3,366,913 op/s » elem
       2,505,998 op/s » elemMod
```